# docs/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "Bank Card Management System API"
  description: "REST API for bank card management with features:\n\n- User authentication\
    \ and authorization\n- Card management (creation, viewing, blocking)\n- Transfers\
    \ between cards\n- System administration\n\n**User Roles:**\n- USER: basic operations\
    \ with own cards\n- ADMIN: full system access\n\n**Security:**\n- Basic Authentication\
    \ (username/password)\n- Card number encryption\n- Sensitive data masking\n"
  contact:
    name: "Abu"
    url: "https://github.com/DesBasito"
    email: "out1of1mind1exception@gmail.com"
  version: "1.0.0"
servers:
- url: "http://localhost:8080/api/v1"
  description: "Local development server"
security:
- Basic Authentication: []
...
paths:
  /api/v1/transactions/{id}/refund:
    put:
      tags:
      - "Transactions"
      summary: "Refund transaction by ID"
      operationId: "toggleTransaction"
      parameters:
      - name: "id"
        in: "path"
        description: "Transaction ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Transaction rolled back successfully!"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "404":
          description: "Transaction not found!"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "403":
          description: "Access denied for this transaction!"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
      security:
      - Basic Authentication: []
...
  /api/v1/transactions/transfer:
    post:
      tags:
      - "Transactions"
      summary: "Transfer between own cards"
      description: "Transfer funds between current user's cards"
      operationId: "transferBetweenMyCards"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
        required: true
      responses:
        "200":
          description: "Transfer completed successfully"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
        "400":
          description: "Invalid transfer data"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "403":
          description: "Card does not belong to user"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
      security:
      - Basic Authentication: []
...
components:
  schemas:
    TransactionDto:
      required:
      - "amount"
      - "createdAt"
      - "status"
      type: "object"
      properties:
        id:
          type: "integer"
          description: "ID транзакции"
          format: "int64"
        fromCardId:
          type: "integer"
          description: "ID карты отправителя"
          format: "int64"
        fromCardNumber:
          type: "string"
          description: "Маскированный номер карты отправителя"
        toCardId:
          type: "integer"
          description: "ID карты получателя"
          format: "int64"
        toCardNumber:
          type: "string"
          description: "Маскированный номер карты получателя"
        amount:
          type: "number"
          description: "Сумма транзакции"
        description:
          type: "string"
          description: "Описание транзакции"
        status:
          type: "string"
          description: "Статус транзакции"
          enum:
          - "SUCCESS"
          - "CANCELLED"
          - "REFUNDED"
        createdAt:
          type: "string"
          description: "Дата создания транзакции"
        processedAt:
          type: "string"
          description: "Дата обработки транзакции"
        errorMessage:
          type: "string"
          description: "Сообщение об ошибке (если есть)"
      description: "Информация о транзакции"
    ApiResponseVoid:
      type: "object"
      properties:
        success:
          type: "boolean"
        message:
          type: "string"
        data:
          type: "object"
    TransferRequest:
      required:
      - "amount"
      - "fromCardId"
      - "toCardId"
      type: "object"
      properties:
        fromCardId:
          type: "integer"
          description: "ID карты отправителя"
          format: "int64"
          example: 1
        toCardId:
          type: "integer"
          description: "ID карты получателя"
          format: "int64"
          example: 2
        amount:
          minimum: 0.01
          exclusiveMinimum: false
          type: "number"
          description: "Сумма перевода"
          example: 1500.5
        description:
          maxLength: 500
          minLength: 0
          type: "string"
          description: "Описание перевода"
          example: "Перевод на другую карту"
      description: "Запрос на перевод между картами"
...
  securitySchemes:
    Basic Authentication:
      type: "http"
      description: "Basic Authentication"
      scheme: "basic"